requirements.txt
audioread==2.1.9
dawdreamer==0.6.10
librosa==0.7.2
music21==7.3.3
openai==0.27.2
Pillow==9.5.0
pydub==0.25.1
pyperclip==1.8.2
Random_Word==1.0.11
requests==2.26.0
scipy==1.7.3
selenium==4.8.3
undetected_chromedriver==3.2.1
